Use the Tile Window from the Shape Tool
I removed the handle to the parent tool from the TileWindow constructor,
as it is unused and prevents reuse.

I have also made the tile window save/load its state to/from the global
editor state, so that the different tile windows behave as one. It might
be neater to go even further, and have the tile window only use the
global editor state instead of storing its own state at all.