Fix warnings
I've chosen not to use abs in the gcd function because it is only
implemented for floats, so ends up implicitly casting. The way that I've
implemented it should work for any possible integer. Yes, even MIN_INT!
(Which is the only int other than 0 that x == -x is true for!)

For the rest, I have added a few more constants to lib.